Tidy up styles
Use grid rather than flex to overlap elements as it avoids a lot of
position absolute.

This in turn lets us let the cards size themselves by just setting an
aspect ratio and either a max width or height (though in practice this
works out similar to scale - it is nicer to set it higher up though and
allows for any number of sizes).

Fan is a pain, the magic numbers for margins annoy me!